hammerhead: update kernel image to fix reboot bug
For a while now, the twrp builds for hammerhead have had an issue
where sometimes it will not boot. After a few seconds kernel panic
and reboot to system. This issue was being caused by some tweaks
in the bricked-kernel currently being used.
https://github.com/showp1984/bricked-hammerhead/commits/twrp

This new kernel image is built from a clean stock source with just
a few changes for twrp.
https://github.com/razorloves/twrp-kernel

Change-Id: I4dd58f7d42b6aae15adf57a4a9c8ff6ef175ba20